Acceleration experiments - C0 for acceleration approximation seems to be working (deceleration + negative velocity handling is missing) Resolution of initial acceleration delta was increased to int32

+++ ControllerCNC/ControllerCNC/DriverCNC.cs
@@ -387,7 +387,7 @@ namespace ControllerCNC
             return deltaT;
         }
 
-        internal void SEND_Acceleration(Int16 stepCount, int initialDeltaT, Int16 n)
+        internal void SEND_Acceleration(Int16 stepCount, int initialDeltaT, int n)
         {
             //Debug.WriteLine("A({0},{1},{2})", stepCount, initialDeltaT, n);
